requirements.txt
selenium
requests
streamlit
pypdf
python-dotenv
beautifulsoup4
Pillow